requirements.txt
biopython==1.79
pandas
einops
matplotlib
requests
pika
pydantic
sentry-sdk